fix selection view
